Add senpatcher_mod.ini support in mod p3a files
This can be used to adjust behavior for a mod that isn't possible via file replacement.
Right now this can only be used to limit a mod file to a specific language (eg. English only), which is useful for Tokyo Xanadu where the English and Japanese script/text files exist at the same path and are switched by whether an extra archive is loaded or not.
In the future this may be extended so that you can adjust in-exe hardcoded values or similar.