Add utility WriteFileAtomic() function
This does not actually do an atomic file write -- that's impossible -- but we can do the next closest thing:
Open tempfile, write to tempfile, rename tempfile to correct filename. On failure, delete tempfile.
This will be observed as atomic for anyone watching that file, though you can see the tempfile if you're looking for it while it's written.